+++ ncf_manager/models/account_invoice.py
@@ -299,16 +299,6 @@ class AccountInvoice(models.Model):
                         })
         return res
 
-    @api.onchange('invoice_line_ids')
-    def _onchange_origin(self):
-        """This method is being inherited as Odoo uses the purchase reference and
-           puts it into the invoice reference (our NCF), we change this behaviour to
-           use the invoice name (description)"""
-        purchase_ids = self.invoice_line_ids.mapped('purchase_id')
-        if purchase_ids:
-            self.origin = ', '.join(purchase_ids.mapped('name'))
-            self.name = ', '.join(purchase_ids.filtered('partner_ref').mapped('partner_ref')) or self.reference
-
     @api.multi
     def invoice_validate(self):
         """ After all invoice validation routine, consume a NCF sequence and write it
